Subject: Add trimming of quotes around branch name returned from API
Description: This change modifies the generate_branch_name function in `gitp.sh` to remove any quotation marks around the branch name returned from the API before returning it. This was necessary to ensure that the branch name is properly formatted and can be used in subsequent git commands. The commit message has been updated to reflect this change.