Subject: Fix issue with adding branch descriptions to Git notes
Description: Previously, the script was not able to add branch descriptions to Git notes due to an issue where the branch description reference was not being checked before attempting to show it. This resulted in errors and prevented the branch description from being appended to the commit message. This commit fixes this issue by adding a check for the existence of the branch description reference before showing it. Additionally, the commit includes a change to the Git notes command syntax to use the branch name in the reference when adding the branch description to the Git notes.