Subject: Add intent functionality to checkout command in gitp script
Description:

Previously, the checkout command in the gitp script did not support specifying an intent for the branch being checked out. This update adds the ability to provide an intent flag when using the checkout command.

The specific changes made include adding a new "intent" variable and setting it to an empty string by default. When the checkout command is used with an intent flag, the value of that flag is assigned to the "intent" variable. This variable is then used to add information to the git log about the purpose of the branch being checked out.

Overall, this change should make it easier for users of the gitp script to track and understand the purpose of their different branches.