Add new flag to edit commit messages
This update adds a new flag (-e or --edit) to allow users to edit commit messages before they are committed. When this flag is set, the Git editor is opened with the current commit message subject and body. If no body exists, only the subject is shown. Once the user saves and exits the editor, the modifications are applied to the commit message. If the commit is being amended, the new message is applied to the previous commit message as well. Otherwise, the new message is used as the commit message. This update also includes minor changes to the install script and the README.