Improve generate_commit_message function to use read instead of array
###RAW###

Improve commit message generator output formatting

The changes made in this commit involve updating the "improve_commit_message" function in "gitp.sh" to improve the output formatting of the generated commit messages. Specifically, the code now uses the "read" command to assign the subject and body of the commit message to variables, which are then combined with the original message to create the final commit message. This change should make the output messages more consistent and easier to read.