Add instructions for committing changes to README
Added instructions to the README on how to stage changes and execute a commit in order to auto-generate a best-practice commit message using gitp. Also modified formatting to improve readability. Additionally, made minor text changes to the uninstallation instructions to clarify the directory that needs to be changed to and updated the chmod command to reference the correct uninstallation script.