Correct spelling of second GPT model in install script
Description: This commit corrects a spelling error in the install.gitp.sh script. The second GPT model option was previously spelled as "gpt4" but has been changed to "gpt-4" to match the naming convention used throughout the script. The change was made on lines 74 and 84 where the model choice values are assigned to the GPT_MODEL_CHOICE environment variable.