Add generate_branch_name and git diff analysis
This commit adds a new function called generate_branch_name to the test.sh file. It also adds a section to analyze the git diff output in the uctags.sh file. This analysis saves the git diff output to a temporary file, reads it, extracts the list of modified functions and variables using perl, removes the '+' and '()' characters to get the item names, and then searches for the definitions of those items in the tags file. Additionally, the global command in uctags.sh is now called with the '-u' option instead of no option, and some minor changes were made to the command used to extract the list of modified functions and variables.