Add support for Ruby, Golang and JavaScript to codequery
This commit adds support for Ruby, Golang, and JavaScript languages to the codequery script. It creates directories for each language and indexes files for each language using starscope tool. Additionally, it removes omnitags and updates the index_files function to account for new languages. The uc2.sh script also has a small bugfix where a comma is added to the end of the languages array.